test(detection): move adapter fakes to shared helpers
